LolEsportsApiClient/requirements.txt
mwrogue
python-dotenv
selenium
beautifulsoup4
